feat: remove reduant controller ans add lioke interactions api
