+++ app/controllers/api/v1/users/current_user_controller.rb
@@ -1,7 +1,17 @@
 module Api
   module V1
     module Users
-      class CurrentUserController < ::Users::CurrentUserController
+      class CurrentUserController < ApplicationController
+        include JwtAuthenticatable
+        
+        before_action :authenticate_user!
+
+        def index
+          render json: {
+            data: UserSerializer.new(current_user).serializable_hash[:data][:attributes],
+            status: :ok
+          }
+        end
       end
     end
   end
