+++ app/models/comment_like.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+class CommentLike < ApplicationRecord
+  belongs_to :comment, counter_cache: :likes_count
+  belongs_to :user
+
+  validates :comment_id, uniqueness: { scope: :user_id }
+end
