+++ app/serializers/comment_serializer.rb
@@ -1,6 +1,19 @@
 class CommentSerializer
   include JSONAPI::Serializer
-  attributes :id, :discussion_id, :user_id, :body, :parent_id, :created_at
+  attributes :id, :discussion_id, :user_id, :body, :parent_id, :created_at, :likes_count
+
+  attribute :is_liked do |comment, params|
+    if params[:current_user]
+      # Check loaded association if available, or query
+      if comment.association(:comment_likes).loaded?
+        comment.comment_likes.any? { |like| like.user_id == params[:current_user].id }
+      else
+        comment.comment_likes.exists?(user_id: params[:current_user].id)
+      end
+    else
+      false
+    end
+  end
 
   attribute :user do |comment|
     {
@@ -11,7 +24,7 @@ class CommentSerializer
     }
   end
 
-  attribute :replies do |comment|
-    CommentSerializer.new(comment.replies).as_json['data']&.map { |d| d['attributes'] } || []
+  attribute :replies do |comment, params|
+    CommentSerializer.new(comment.replies, { params: params }).serializable_hash[:data]&.map { |d| d[:attributes] } || []
   end
 end
